+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix bug where --non-interactive was inversed
 - `./result` links where being created. they will not be created anymore
+- Logging from interactive commands (absence of `--non-interactive`) was
+  improved.
 
 ### Changed